Added int64 pointers (#165)
* using long long pointers

* added ifdef for linux for stdint

* version bump to 0.39.8

+++ pyansys/cython/binary_reader.h
@@ -1,5 +1,5 @@
 /* #include <fstream> */
 
-void read_nodes(const char*, int, int, int *, double *);
-void* read_record(const char*, int, int*, int*, int*, int*);
-void read_record_stream(std::ifstream*, int, void*, int*, int*, int*);
+void read_nodes(const char*, int64_t, int, int *, double *);
+void* read_record(const char*, int64_t, int*, int*, int*, int*);
+void read_record_stream(std::ifstream*, int64_t, void*, int*, int*, int*);
